+++ src/Parser/Priest/Holy/Modules/Talents/15/EnduringRenewal.js
@@ -6,7 +6,7 @@ import { formatPercentage, formatNumber } from 'common/format';
 import SPELLS from 'common/SPELLS/index';
 import ITEMS from 'common/ITEMS/index';
 import Analyzer from 'Parser/Core/Analyzer';
-import { ABILITIES_THAT_TRIGGER_ENDURING_RENEWAL } from '../../Constants';
+import { ABILITIES_THAT_TRIGGER_ENDURING_RENEWAL } from '../../../Constants';
 
 class EnduringRenewal extends Analyzer {
   _normalRenewDropoff = {};
@@ -72,36 +72,6 @@ class EnduringRenewal extends Analyzer {
       this._newRenewDropoff[event.targetID] = event.timestamp + this._baseRenewLength * 1000;
     }
   }
-
-
-  statistic() {
-    const erPercHPS = formatPercentage(this.owner.getPercentageOfTotalHealingDone(this.healing));
-    const erHPS = formatNumber(this.healing / this.owner.fightDuration * 1000);
-    const erGainPerRefresh = Math.round(this.secsGained / this.refreshedRenews * 100) / 100;
-
-    //
-    return this.active && (
-      <StatisticBox
-        icon={<SpellIcon id={SPELLS.ENDURING_RENEWAL_TALENT.id} />}
-        value={`${erHPS} HPS`}
-        label={(
-          <dfn data-tip={`
-            Healing done on targets as a result of Enduring Renewal's refresh.
-            This did ${formatNumber(this.healing)} healing and was ${erPercHPS}% of your total healing.
-            <br/><br/>
-            You refreshed renews ${this.refreshedRenews} times for a total of ${formatNumber(this.secsGained)} additional seconds of Renew.
-            (+${erGainPerRefresh}s per refresh on average).
-          `}
-          >
-            Enduring Renewal
-          </dfn>
-        )}
-      />
-    );
-    //
-  }
-
-  statisticOrder = STATISTIC_ORDER.OPTIONAL(1);
 }
